+++ app/code/community/Ecocode/Profiler/Model/Collector/AbstractDataCollector.php
@@ -9,6 +9,7 @@ abstract class Ecocode_Profiler_Model_Collector_AbstractDataCollector
 {
     protected $data = [];
 
+    protected $helper;
     protected $contextHelper;
 
     public function init()
@@ -58,6 +59,7 @@ abstract class Ecocode_Profiler_Model_Collector_AbstractDataCollector
     }
 
     /**
+     * @codeCoverageIgnore
      * @return Ecocode_Profiler_Helper_Context
      */
     public function getContextHelper()
@@ -69,6 +71,23 @@ abstract class Ecocode_Profiler_Model_Collector_AbstractDataCollector
         return $this->contextHelper;
     }
 
+    /**
+     * @codeCoverageIgnore
+     * @return Ecocode_Profiler_Helper_Data
+     */
+    protected function getHelper()
+    {
+        if ($this->helper === null) {
+            $this->helper = Mage::helper('ecocode_profiler');
+        }
+        return $this->helper;
+    }
+
+    /**
+     * @codeCoverageIgnore
+     * @param int $options
+     * @return array|bool
+     */
     protected function getBacktrace($options = DEBUG_BACKTRACE_PROVIDE_OBJECT)
     {
         if (!function_exists('debug_backtrace')) {