+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 CHANGELOG
 =========
 
-next version
+1.2.0
 -----
 * fixed an exception in debug mode when block output is suppressed with "disable module output"
-* fixed an exception when logging and monolog is not installed
+* fixed an exception if monolog is not installed
 * unified call stack rendering
-* improved request collector to capture also redirects
+* improved request collector to capture also redirects, session data and flash messages
 * improved support for "xdebug.file_link_format" to also work with urls like "http://localhost:63342/api/file/%f:%l"
 
 1.1.0